get log in to function
